pyproject.toml
[tool.poetry]
name = "filonna"
version = "0.1.0"
description = "A CLI based movie metadata and folder manager"
authors = ["sachin-sankar <mail.sachinsankar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
platformdirs = "^4.3.6"
rich = "^13.9.3"
tmdbsimple = "^2.9.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
